requirements.txt
Django==5.2.11
mssql-django==1.6
pyjwt==2.11.0

# The following are indirect dependencies but pinned here for reproducibility
asgiref==3.11.1
pyodbc==5.3.0
pytz==2025.2
sqlparse==0.5.5

Django==5.2.11
mssql-django==1.6
pyjwt==2.11.0
pyodbc==5.3.0
